use typed Const16 in binary wasmi instructions

+++ crates/wasmi/src/engine/tests/regmach/op/binary/i32_rem_s.rs
@@ -18,12 +18,12 @@ fn reg_reg() {
 
 #[test]
 fn reg_imm16() {
-    test_binary_reg_imm16::<i16>(WASM_OP, 100, Instruction::i32_rem_s_imm16)
+    test_binary_reg_imm16::<i32>(WASM_OP, 100, Instruction::i32_rem_s_imm16)
 }
 
 #[test]
 fn reg_imm16_rev() {
-    test_binary_reg_imm16_rev::<i16>(WASM_OP, 100, Instruction::i32_rem_s_imm16_rev)
+    test_binary_reg_imm16_rev::<i32>(WASM_OP, 100, Instruction::i32_rem_s_imm16_rev)
 }
 
 #[test]
