+++ crates/wasmi/src/engine/tests/regmach/op/cmp/i64_gt_u.rs
@@ -17,12 +17,12 @@ fn reg_reg() {
 
 #[test]
 fn reg_imm16() {
-    test_binary_reg_imm16::<u16>(WASM_OP, 100, Instruction::i64_gt_u_imm16)
+    test_binary_reg_imm16::<u64>(WASM_OP, 100, Instruction::i64_gt_u_imm16)
 }
 
 #[test]
 fn reg_imm16_rev() {
-    test_binary_reg_imm16_rev::<u16>(WASM_OP, 100, swap_ops!(Instruction::i64_lt_u_imm16))
+    test_binary_reg_imm16_rev::<u64>(WASM_OP, 100, swap_ops!(Instruction::i64_lt_u_imm16))
 }
 
 #[test]
